refactor: split tu_merge.py/qualified_name_segments.py to stay under the architecture size gates
The two prior fix commits (source_header_fact/qualified_name_fact sync)
pushed tu_merge.py past its ADR-061 no-growth debt baseline (1552 lines)
and qualified_name_segments.py past the 800-line production cap, plus
their test files past their own caps -- confirmed against CI's real
architecture job failure on aa1e3bf.

- abicheck/tu_merge_provenance.py (new): tu_merge.py's provenance-
  comparison cluster (_pick_deprecated, _blank_provenance,
  _more_public_of, _other_is_strictly_less_public,
  _with_more_public_provenance) -- mechanical extraction, unchanged
  bodies, re-imported so no external call site changes. tu_merge.py:
  1580 -> 1317 lines.
- abicheck/qualified_name_segments_walk.py (new): the generic payload-
  exclusion-aware dataclass string walk (_PAYLOAD_FIELD_EXCLUSIONS,
  _collect_strings, _legacy_sibling_is_payload_excluded,
  _walk_rewrite_strings) -- same mechanical split.
  qualified_name_segments.py: 856 -> 682 lines.
- architecture/modules.yaml: registers both new root modules in
  legacy_root_modules (the check-architecture.py "root-module" gate).
- tests/test_tu_merge_fact_provenance.py,
  tests/test_lambda_identity_payload_exclusions.py (new): the fact-sync
  regression tests split out to keep their own parent test files under
  the debt baseline / test-file cap, alongside their fixtures.

Every function body is unchanged; this is purely a location split to
satisfy scripts/check_architecture.py's debt-no-growth and file-size
gates, verified via `python scripts/verify.py --profile pr --only
architecture` plus the full targeted test suite.
