+++ architecture/modules.yaml
@@ -822,6 +822,7 @@
     "python_api.py",
     "python_ext.py",
     "qualified_name_segments.py",
+    "qualified_name_segments_walk.py",
     "reclassify.py",
     "report_classifications.py",
     "report_correlation.py",
@@ -868,6 +869,7 @@
     "symvers_metadata.py",
     "tu_fragment.py",
     "tu_merge.py",
+    "tu_merge_provenance.py",
     "type_metadata.py",
     "type_reachability.py",
     "type_reachability_spelling.py",
